sepolia/pyton-deployer/requirements.txt
web3
py-solc-x
python-dotenv
